requirements.txt
fastapi~=0.115.13
uvicorn
pymongo~=4.13.2
qdrant-client~=1.14.3
python-multipart
pymupdf
python-docx~=1.2.0
requests~=2.32.4
sentence-transformers~=4.1.0

numpy~=2.3.1
scikit-learn~=1.7.0

pdfplumber